use paths from asset files for editor styles
