+++ wp-content/plugins/core/src/Theme/Theme_Subscriber.php
@@ -7,8 +7,6 @@ use Tribe\Libs\Container\Abstract_Subscriber;
 use Tribe\Project\Theme\Config\Image_Sizes;
 use Tribe\Project\Theme\Config\Supports;
 use Tribe\Project\Theme\Config\Web_Fonts;
-use Tribe\Project\Theme\Editor\Classic_Editor_Formats;
-use Tribe\Project\Theme\Editor\Editor_Styles;
 use Tribe\Project\Theme\Media\Full_Size_Gif;
 use Tribe\Project\Theme\Media\Image_Wrap;
 use Tribe\Project\Theme\Media\Oembed_Filter;
@@ -19,7 +17,6 @@ class Theme_Subscriber extends Abstract_Subscriber {
 		$this->body_classes();
 		$this->media();
 		$this->config();
-		$this->editor();
 	}
 
 	private function config(): void {
@@ -33,7 +30,7 @@ class Theme_Subscriber extends Abstract_Subscriber {
 		$this->image_links();
 		$this->disable_responsive_images();
 		$this->oembed();
-		// $this->full_size_gif(); Uncomment to require full size gifs
+		// $this->full_size_gif(); // Uncomment to require full size gifs
 	}
 
 	private function body_classes() {
@@ -114,32 +111,6 @@ class Theme_Subscriber extends Abstract_Subscriber {
 		/* add_action( 'login_head', function() {
 			$this->container->get( Fonts::class )->load_fonts();
 		}, 0, 0); */
-	}
-
-	private function editor(): void {
-		$this->editor_styles();
-		//$this->editor_formats();
-	}
-
-	private function editor_styles() {
-		add_action( 'admin_init', function () {
-			$this->container->get( Editor_Styles::class )->block_editor_styles();
-		}, 10, 0 );
-		add_filter( 'tiny_mce_before_init', function ( $settings ) {
-			return $this->container->get( Editor_Styles::class )->mce_editor_body_class( $settings );
-		}, 10, 1 );
-		add_filter( 'editor_stylesheets', function ( $styles ) {
-			return $this->container->get( Editor_Styles::class )->mce_editor_styles( $styles );
-		}, 10, 1 );
-	}
-
-	private function editor_formats() {
-		add_filter( 'mce_buttons', function ( $settings ) {
-			return $this->container->get( Classic_Editor_Formats::class )->mce_buttons( $settings );
-		}, 10, 1 );
-		add_filter( 'tiny_mce_before_init', function ( $settings ) {
-			return $this->container->get( Classic_Editor_Formats::class )->visual_editor_styles_dropdown( $settings );
-		}, 10, 1 );
 	}
 
 }
